+++ CHANGES.md
@@ -1,3 +1,11 @@
+#### [Version 0.3.0 Beta](https://github.com/Avaq/JSLite/tree/0.3.0-beta)
+_03-Sep-2013_
+
+* Classes gain some standard properties from ClassFactory.
+  - `classInstance._STATIC` will always refer to the constructor of the instance.
+  - `classInstance.proxy()` is available on every instance through `Class.prototype`.
+* Updated the minified version.
+
 #### [Version 0.2.0 Beta](https://github.com/Avaq/JSLite/tree/0.2.0-beta)
 _18-Jul-2013_